+++ lib/bloc_notification.dart
@@ -1,4 +1,6 @@
+export 'src/bloc_notification.dart';
 export 'src/bloc_notification_consumer.dart';
 export 'src/bloc_notification_listener.dart';
 export 'src/bloc_notification_mixin.dart';
+export 'src/notification_observer.dart';
 export 'src/type.dart';