+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* No child bug fixed in `BlocNotificationListener`
+* Remove dependency on `rxdart`
+* Additional callback `onNotification` inside `BlocNotificationMixins`
+* Mixin for `BlocObserver` to listen to notification
+* More documentations
+
 ## 1.0.0+1
 
 * Add badge in README